fix: remove unnecessary u32 cast in cranelift codegen
cl_ty.bytes() already returns u32; the cast was redundant and
flagged by clippy::unnecessary_cast on Rust 1.95.

+++ compiler/crates/kryos-codegen-cranelift/src/codegen.rs
@@ -213,7 +213,7 @@ fn compute_struct_layout(fields: &[(String, MirType)]) -> Result<StructLayout, C
     let mut field_offsets = Vec::new();
     for (name, ty) in fields {
         let cl_ty = mir_type_to_cl(ty)?.unwrap_or(types::I64);
-        let size = cl_ty.bytes() as u32;
+        let size = cl_ty.bytes();
         // Natural alignment: align to the field's own size.
         let align = size;
         if align > 0 {
